Prevent negative like count when disliking music

+++ backend/src/modules/music/useCases/DeslikeMusic.ts
@@ -18,10 +18,14 @@ class DeslikeMusic {
         const interactedMusics = await this.musicActionRepository.getInteractedMusicsByUser(user.id);
         if((interactedMusics.find(music=>music.musicId===musicId && music.reaction==="DISLIKE")))throw new Error("Music already desliked by user");
         
-        await this.musicRepository.removeLike(musicId);
+        if(musicExists.likeCount > 0){
+            await this.musicRepository.removeLike(musicId)
+        };
+        
+        await this.musicActionRepository.upsert({musicId,userId:user.id,reaction:"DISLIKE"});
+
         
 
-        await this.musicActionRepository.upsert({musicId,userId:user.id,reaction:"DISLIKE"});
         
     }
 }
